introduce sparse_labels to cpp

+++ predict.py
@@ -50,7 +50,7 @@ class Predictor:
             # Graph for interpolating labels
             # Assuming batch_size == 1 for simplicity
             pl_sparse_points = tf.placeholder(tf.float32, (None, 3))
-            pl_sparse_labels = tf.placeholder(tf.float32, (None,))
+            pl_sparse_labels = tf.placeholder(tf.int32, (None,))
             pl_dense_points = tf.placeholder(tf.float32, (None, 3))
             _, sparse_indices = interpolate_label(
                 pl_sparse_points, pl_sparse_labels, pl_dense_points
